Replace slide link with Twitter check-in
Instead of pointing attendees just to the slide deck, direct them to
Twitter where they can let the world know they're in the talk. Also,
include the link to the slides in the Twitter message.

Uses Tweet Web Intents
(https://dev.twitter.com/web/tweet-button/web-intent) in combination
with a URL encoder and goo.gl (to simplify the link, which in turn
makes the QR code scan faster).